Replaced normals type from float to normalized int8_t

+++ include/icosphere.h
@@ -32,7 +32,7 @@ class Icosphere {
   std::vector<Triangle*> triangles_;
 
   float* vertices_array_;
-  float* normals_array_;
+  int8_t* normals_array_;
   uint8_t* colors_array_;
   uint16_t* indices_array_;
   uint16_t* tex_coord_array_;
